+++ framework/modelcatalog/datasheet/overrides_test.go
@@ -1,4 +1,4 @@
-package modelcatalog
+package datasheet
 
 import (
 	"testing"
@@ -23,10 +23,22 @@ func (noOpLogger) LogHTTPRequest(schemas.LogLevel, string) schemas.LogEventBuild
 	return schemas.NoopLogEvent
 }
 
+// newTestStore builds a minimal Store for unit tests. Callers seed pricingData
+// directly and use SetOverrides for overrides.
+func newTestStore() *Store {
+	return &Store{
+		logger:                 noOpLogger{},
+		pricingData:            map[string]configstoreTables.TableModelPricing{},
+		baseModelIndex:         map[string]string{},
+		supportedResponseTypes: map[string][]string{},
+		supportedParams:        map[string][]string{},
+		datasheetByProvider:    map[schemas.ModelProvider][]string{},
+	}
+}
+
 func TestGetPricing_OverridePrecedenceExactWildcard(t *testing.T) {
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o",
 		Provider:           "openai",
 		Mode:               "chat",
@@ -35,7 +47,7 @@ func TestGetPricing_OverridePrecedenceExactWildcard(t *testing.T) {
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "openai-override-0",
 			ScopeKind:        string(ScopeKindProvider),
@@ -56,7 +68,7 @@ func TestGetPricing_OverridePrecedenceExactWildcard(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	require.NotNil(t, pricing.InputCostPerToken)
 	assert.Equal(t, 20.0, *pricing.InputCostPerToken)
@@ -64,9 +76,8 @@ func TestGetPricing_OverridePrecedenceExactWildcard(t *testing.T) {
 
 func TestGetPricing_RequestTypeSpecificOverrideBeatsGeneric(t *testing.T) {
 	t.Skip()
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o", "openai", "responses")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o", "openai", "responses")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o",
 		Provider:           "openai",
 		Mode:               "responses",
@@ -75,7 +86,7 @@ func TestGetPricing_RequestTypeSpecificOverrideBeatsGeneric(t *testing.T) {
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "openai-generic",
 			ScopeKind:        string(ScopeKindProvider),
@@ -95,16 +106,15 @@ func TestGetPricing_RequestTypeSpecificOverrideBeatsGeneric(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ResponsesRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ResponsesRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 15.0, pricing.InputCostPerToken)
 }
 
 func TestGetPricing_AppliesOverrideAfterFallbackResolution(t *testing.T) {
 	t.Skip()
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o", "vertex", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o", "vertex", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o",
 		Provider:           "vertex",
 		Mode:               "chat",
@@ -113,7 +123,7 @@ func TestGetPricing_AppliesOverrideAfterFallbackResolution(t *testing.T) {
 	}
 
 	geminiProviderID := "gemini"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "gemini-override",
 			ScopeKind:        string(ScopeKindProvider),
@@ -124,15 +134,14 @@ func TestGetPricing_AppliesOverrideAfterFallbackResolution(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "gemini", Model: "gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "gemini"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "gemini", Model: "gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "gemini"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 7.0, pricing.InputCostPerToken)
 }
 
 func TestGetPricing_DeploymentLookupUsesResolvedModelForOverrideMatching(t *testing.T) {
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("dep-gpt4o", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("dep-gpt4o", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "dep-gpt4o",
 		Provider:           "openai",
 		Mode:               "chat",
@@ -141,7 +150,7 @@ func TestGetPricing_DeploymentLookupUsesResolvedModelForOverrideMatching(t *test
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "resolved-model-override",
 			ScopeKind:        string(ScopeKindProvider),
@@ -155,16 +164,15 @@ func TestGetPricing_DeploymentLookupUsesResolvedModelForOverrideMatching(t *test
 
 	// Override pattern matches the resolved model name ("dep-gpt4o"), not the
 	// originally requested name ("gpt-4o"), because resolved model has priority.
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o", ResolvedKeyAlias: &schemas.ResolvedKeyAlias{ModelID: "dep-gpt4o"}}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o", ResolvedKeyAlias: &schemas.ResolvedKeyAlias{ModelID: "dep-gpt4o"}}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	require.NotNil(t, pricing.InputCostPerToken)
 	assert.Equal(t, 7.0, *pricing.InputCostPerToken)
 }
 
 func TestGetPricing_FallbackUsesRequestedProviderForScopeMatching(t *testing.T) {
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o", "vertex", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o", "vertex", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o",
 		Provider:           "vertex",
 		Mode:               "chat",
@@ -174,7 +182,7 @@ func TestGetPricing_FallbackUsesRequestedProviderForScopeMatching(t *testing.T) 
 
 	geminiProviderID := "gemini"
 	vertexProviderID := "vertex"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "gemini-provider-override",
 			ScopeKind:        string(ScopeKindProvider),
@@ -195,7 +203,7 @@ func TestGetPricing_FallbackUsesRequestedProviderForScopeMatching(t *testing.T) 
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "gemini", Model: "gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "gemini"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "gemini", Model: "gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "gemini"})
 	require.NotNil(t, pricing)
 	require.NotNil(t, pricing.InputCostPerToken)
 	assert.Equal(t, 5.0, *pricing.InputCostPerToken)
@@ -203,9 +211,8 @@ func TestGetPricing_FallbackUsesRequestedProviderForScopeMatching(t *testing.T) 
 
 func TestGetPricing_ExactOverrideDoesNotMatchProviderPrefixedModel(t *testing.T) {
 	t.Skip()
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("openai/gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("openai/gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "openai/gpt-4o",
 		Provider:           "openai",
 		Mode:               "chat",
@@ -214,7 +221,7 @@ func TestGetPricing_ExactOverrideDoesNotMatchProviderPrefixedModel(t *testing.T)
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "openai-override-0",
 			ScopeKind:        string(ScopeKindProvider),
@@ -225,17 +232,16 @@ func TestGetPricing_ExactOverrideDoesNotMatchProviderPrefixedModel(t *testing.T)
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "openai/gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "openai/gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 1.0, pricing.InputCostPerToken)
 }
 
 func TestGetPricing_NoMatchingOverrideLeavesPricingUnchanged(t *testing.T) {
 	t.Skip()
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
+	s := newTestStore()
 	baseCacheRead := 0.4
-	mc.pricingData[makeKey("gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s.pricingData[makeKey("gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:                   "gpt-4o",
 		Provider:                "openai",
 		Mode:                    "chat",
@@ -245,7 +251,7 @@ func TestGetPricing_NoMatchingOverrideLeavesPricingUnchanged(t *testing.T) {
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "openai-override-0",
 			ScopeKind:        string(ScopeKindProvider),
@@ -256,7 +262,7 @@ func TestGetPricing_NoMatchingOverrideLeavesPricingUnchanged(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 1.0, pricing.InputCostPerToken)
 	assert.Equal(t, 2.0, pricing.OutputCostPerToken)
@@ -264,11 +270,10 @@ func TestGetPricing_NoMatchingOverrideLeavesPricingUnchanged(t *testing.T) {
 	assert.Equal(t, 0.4, *pricing.CacheReadInputTokenCost)
 }
 
-func TestDeleteProviderPricingOverrides_StopsApplying(t *testing.T) {
+func TestDeleteProviderOverrides_StopsApplying(t *testing.T) {
 	t.Skip()
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o",
 		Provider:           "openai",
 		Mode:               "chat",
@@ -277,7 +282,7 @@ func TestDeleteProviderPricingOverrides_StopsApplying(t *testing.T) {
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "openai-override-0",
 			ScopeKind:        string(ScopeKindProvider),
@@ -288,22 +293,21 @@ func TestDeleteProviderPricingOverrides_StopsApplying(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 11.0, pricing.InputCostPerToken)
 
-	require.NoError(t, mc.SetPricingOverrides(nil))
+	require.NoError(t, s.SetOverrides(nil))
 
-	pricing = mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing = s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 1.0, pricing.InputCostPerToken)
 }
 
 func TestGetPricing_WildcardSpecificityLongerLiteralWins(t *testing.T) {
 	t.Skip()
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o-mini", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o-mini", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o-mini",
 		Provider:           "openai",
 		Mode:               "chat",
@@ -312,7 +316,7 @@ func TestGetPricing_WildcardSpecificityLongerLiteralWins(t *testing.T) {
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "openai-override-0",
 			ScopeKind:        string(ScopeKindProvider),
@@ -331,17 +335,14 @@ func TestGetPricing_WildcardSpecificityLongerLiteralWins(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o-mini"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o-mini"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	assert.Equal(t, 6.0, pricing.InputCostPerToken)
 }
 
-// TestGetPricing_FirstInsertionWinsOnTie verifies that when multiple wildcard overrides
-// match the same model and scope, the first one inserted takes precedence.
 func TestGetPricing_FirstInsertionWinsOnTie(t *testing.T) {
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
-	mc.pricingData[makeKey("gpt-4o-mini", "openai", "chat")] = configstoreTables.TableModelPricing{
+	s := newTestStore()
+	s.pricingData[makeKey("gpt-4o-mini", "openai", "chat")] = configstoreTables.TableModelPricing{
 		Model:              "gpt-4o-mini",
 		Provider:           "openai",
 		Mode:               "chat",
@@ -350,7 +351,7 @@ func TestGetPricing_FirstInsertionWinsOnTie(t *testing.T) {
 	}
 
 	providerID := "openai"
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "a-override",
 			ScopeKind:        string(ScopeKindProvider),
@@ -371,7 +372,7 @@ func TestGetPricing_FirstInsertionWinsOnTie(t *testing.T) {
 		},
 	}))
 
-	pricing := mc.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o-mini"}, schemas.ChatCompletionRequest, PricingLookupScopes{Provider: "openai"})
+	pricing := s.resolvePricing(schemas.RoutingInfo{Provider: "openai", Model: "gpt-4o-mini"}, schemas.ChatCompletionRequest, LookupScopes{Provider: "openai"})
 	require.NotNil(t, pricing)
 	require.NotNil(t, pricing.InputCostPerToken)
 	assert.Equal(t, 8.0, *pricing.InputCostPerToken)
@@ -392,7 +393,7 @@ func TestPatchPricing_PartialPatchOnlyChangesSpecifiedFields(t *testing.T) {
 	}
 
 	cacheRead := 0.9
-	patched := patchPricing(base, PricingOptions{
+	patched := patchPricing(base, Options{
 		InputCostPerToken:       bifrost.Ptr(3.0),
 		CacheReadInputTokenCost: &cacheRead,
 	})
@@ -406,15 +407,14 @@ func TestPatchPricing_PartialPatchOnlyChangesSpecifiedFields(t *testing.T) {
 	assert.Equal(t, 0.7, *patched.InputCostPerImage)
 }
 
-func TestApplyScopedPricingOverrides_ScopePrecedence(t *testing.T) {
-	mc := newTestCatalog(nil, nil)
-	mc.logger = noOpLogger{}
+func TestApplyScopedOverrides_ScopePrecedence(t *testing.T) {
+	s := newTestStore()
 
 	providerScopeID := "openai"
 	providerKeyScopeID := "provider-key-1"
 	virtualKeyScopeID := "virtual-key-1"
 
-	require.NoError(t, mc.SetPricingOverrides([]configstoreTables.TablePricingOverride{
+	require.NoError(t, s.SetOverrides([]configstoreTables.TablePricingOverride{
 		{
 			ID:               "global",
 			ScopeKind:        string(ScopeKindGlobal),
@@ -462,12 +462,12 @@ func TestApplyScopedPricingOverrides_ScopePrecedence(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		scopes   PricingLookupScopes
+		scopes   LookupScopes
 		expected float64
 	}{
 		{
 			name: "virtual key wins over provider key, provider and global",
-			scopes: PricingLookupScopes{
+			scopes: LookupScopes{
 				VirtualKeyID:  virtualKeyScopeID,
 				SelectedKeyID: providerKeyScopeID,
 				Provider:      providerScopeID,
@@ -476,7 +476,7 @@ func TestApplyScopedPricingOverrides_ScopePrecedence(t *testing.T) {
 		},
 		{
 			name: "provider key wins over provider and global",
-			scopes: PricingLookupScopes{
+			scopes: LookupScopes{
 				SelectedKeyID: providerKeyScopeID,
 				Provider:      providerScopeID,
 			},
@@ -484,21 +484,21 @@ func TestApplyScopedPricingOverrides_ScopePrecedence(t *testing.T) {
 		},
 		{
 			name: "provider wins over global",
-			scopes: PricingLookupScopes{
+			scopes: LookupScopes{
 				Provider: providerScopeID,
 			},
 			expected: 3.0,
 		},
 		{
 			name:     "global applies when no narrower scope is provided",
-			scopes:   PricingLookupScopes{},
+			scopes:   LookupScopes{},
 			expected: 2.0,
 		},
 	}
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			patched, applied := mc.applyPricingOverrides("gpt-5-nano", schemas.ChatCompletionRequest, base, tc.scopes)
+			patched, applied := s.applyPricingOverrides("gpt-5-nano", schemas.ChatCompletionRequest, base, tc.scopes)
 			require.True(t, applied)
 			require.NotNil(t, patched.InputCostPerToken)
 			assert.Equal(t, tc.expected, *patched.InputCostPerToken)
