+++ framework/modelcatalog/models.go
@@ -7,228 +7,225 @@ import (
 
 	"github.com/maximhq/bifrost/core/schemas"
 	"github.com/maximhq/bifrost/framework/configstore"
-	configstoreTables "github.com/maximhq/bifrost/framework/configstore/tables"
 )
 
-// GetModelCapabilityEntryForModel returns capability metadata for a model/provider pair.
-// It prefers chat, then responses, then text-completion entries; if none exist,
-// it falls back to the lexicographically first available mode for deterministic behavior.
-func (mc *ModelCatalog) GetModelCapabilityEntryForModel(model string, provider schemas.ModelProvider) *PricingEntry {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
+// GetModelsForProvider returns the effective allowed model set for the
+// provider. Filtered live entries are authoritative when present (they were
+// pre-gated by ListModelsPipeline against the key's allow/block/aliases);
+// otherwise the datasheet view is filtered by the keyconfig aggregates.
+func (mc *ModelCatalog) GetModelsForProvider(provider schemas.ModelProvider) []string {
+	blacklisted := mc.keyconf.BlacklistedFor(provider)
+	allowed := mc.keyconf.AllowedFor(provider)
 
-	if entry := mc.getCapabilityEntryForExactModelUnsafe(model, provider); entry != nil {
-		return entry
+	var out []string
+	if liveModels := mc.live.ModelsForProvider(provider); len(liveModels) > 0 {
+		out = liveModels
+	} else if datasheetModels := mc.datasheet.DatasheetModelsForProvider(provider); len(datasheetModels) > 0 && allowed != nil {
+		out = make([]string, 0, len(datasheetModels))
+		for _, m := range datasheetModels {
+			if blacklisted.IsBlocked(m) {
+				continue
+			}
+			if allowed.IsAllowed(m) {
+				out = append(out, m)
+			}
+		}
+	} else {
+		out = []string{}
 	}
 
-	baseModel := mc.getBaseModelNameUnsafe(model)
-	if baseModel != model {
-		if entry := mc.getCapabilityEntryForExactModelUnsafe(baseModel, provider); entry != nil {
-			return entry
+	seen := make(map[string]struct{}, len(out))
+	for _, m := range out {
+		seen[m] = struct{}{}
+	}
+	for _, e := range mc.keyconf.EntriesFor(provider) {
+		if !e.Enabled {
+			continue
+		}
+		for alias := range e.Aliases {
+			if blacklisted.IsBlocked(alias) {
+				continue
+			}
+			if allowed == nil || !allowed.IsAllowed(alias) {
+				continue
+			}
+			if _, ok := seen[alias]; ok {
+				continue
+			}
+			seen[alias] = struct{}{}
+			out = append(out, alias)
+		}
+		for _, m := range e.Allowed {
+			if m == "*" || blacklisted.IsBlocked(m) {
+				continue
+			}
+			if _, ok := seen[m]; ok {
+				continue
+			}
+			seen[m] = struct{}{}
+			out = append(out, m)
 		}
 	}
-
-	if entry := mc.getCapabilityEntryForModelFamilyUnsafe(baseModel, provider); entry != nil {
-		return entry
-	}
-
-	return nil
+	return out
 }
 
-// GetModelsForProvider returns all available models for a given provider (thread-safe)
-func (mc *ModelCatalog) GetModelsForProvider(provider schemas.ModelProvider) []string {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
-
-	models, exists := mc.modelPool[provider]
-	if !exists {
-		return []string{}
-	}
-
-	// Return a copy to prevent external modification
-	result := make([]string, len(models))
-	copy(result, models)
-	return result
-}
-
-// GetUnfilteredModelsForProvider returns all available models for a given provider (thread-safe)
+// GetUnfilteredModelsForProvider returns the raw catalog view (no gate
+// applied): union of live unfiltered entries and the datasheet view.
 func (mc *ModelCatalog) GetUnfilteredModelsForProvider(provider schemas.ModelProvider) []string {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
-
-	models, exists := mc.unfilteredModelPool[provider]
-	if !exists {
-		return []string{}
+	liveModels := mc.live.UnfilteredModelsForProvider(provider)
+	datasheetModels := mc.datasheet.DatasheetModelsForProvider(provider)
+	if len(liveModels) == 0 {
+		return datasheetModels
 	}
-
-	// Return a copy to prevent external modification
-	result := make([]string, len(models))
-	copy(result, models)
-	return result
+	if len(datasheetModels) == 0 {
+		return liveModels
+	}
+	seen := make(map[string]struct{}, len(liveModels)+len(datasheetModels))
+	out := make([]string, 0, len(liveModels)+len(datasheetModels))
+	for _, m := range liveModels {
+		if _, ok := seen[m]; !ok {
+			seen[m] = struct{}{}
+			out = append(out, m)
+		}
+	}
+	for _, m := range datasheetModels {
+		if _, ok := seen[m]; !ok {
+			seen[m] = struct{}{}
+			out = append(out, m)
+		}
+	}
+	slices.Sort(out)
+	return out
 }
 
-// GetDistinctBaseModelNames returns all unique base model names from the catalog (thread-safe).
-// This is used for governance model selection when no specific provider is chosen.
+// GetDistinctBaseModelNames returns all unique base model names from the
+// datasheet. Used by governance for cross-provider model selection.
 func (mc *ModelCatalog) GetDistinctBaseModelNames() []string {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
-
-	seen := make(map[string]bool)
-	for _, baseName := range mc.baseModelIndex {
-		seen[baseName] = true
-	}
-
-	result := make([]string, 0, len(seen))
-	for name := range seen {
-		result = append(result, name)
-	}
-	return result
+	return mc.datasheet.DistinctBaseModelNames()
 }
 
-// GetProvidersForModel returns all providers for a given model (thread-safe)
+// GetProvidersForModel returns every provider that can serve the model.
+// Composes across stores and applies the cross-provider special cases
+// (openrouter / vertex / groq-gpt / bedrock-claude) preserved verbatim from
+// the pre-refactor implementation.
 func (mc *ModelCatalog) GetProvidersForModel(model string) []schemas.ModelProvider {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
+	baseModel := mc.datasheet.BaseModelName(model)
 
 	providers := make([]schemas.ModelProvider, 0)
-	for provider, models := range mc.modelPool {
-		isModelMatch := false
+	seen := make(map[schemas.ModelProvider]struct{})
+	for _, p := range mc.knownProviders() {
+		models := mc.GetModelsForProvider(p)
+		matched := false
 		for _, m := range models {
-			if m == model || mc.getBaseModelNameUnsafe(m) == mc.getBaseModelNameUnsafe(model) {
-				isModelMatch = true
+			if m == model || mc.datasheet.BaseModelName(m) == baseModel {
+				matched = true
 				break
 			}
 		}
-		if isModelMatch {
-			providers = append(providers, provider)
-		}
-	}
-
-	// Handler special provider cases
-	// 1. Handler openrouter models
-	if !slices.Contains(providers, schemas.OpenRouter) {
-		for _, provider := range providers {
-			if openRouterModels, ok := mc.modelPool[schemas.OpenRouter]; ok {
-				if slices.Contains(openRouterModels, string(provider)+"/"+model) {
-					providers = append(providers, schemas.OpenRouter)
-				}
+		if matched {
+			if _, ok := seen[p]; !ok {
+				providers = append(providers, p)
+				seen[p] = struct{}{}
 			}
 		}
 	}
 
-	// 2. Handle vertex models
-	if !slices.Contains(providers, schemas.Vertex) {
-		for _, provider := range providers {
-			if vertexModels, ok := mc.modelPool[schemas.Vertex]; ok {
-				if slices.Contains(vertexModels, string(provider)+"/"+model) {
-					providers = append(providers, schemas.Vertex)
-				}
+	// Cross-provider special cases (preserved from pre-refactor models.go).
+	if _, ok := seen[schemas.OpenRouter]; !ok {
+		openRouterModels := mc.GetModelsForProvider(schemas.OpenRouter)
+		for _, p := range providers {
+			if slices.Contains(openRouterModels, string(p)+"/"+model) {
+				providers = append(providers, schemas.OpenRouter)
+				seen[schemas.OpenRouter] = struct{}{}
+				break
+			}
+		}
+	}
+	if _, ok := seen[schemas.Vertex]; !ok {
+		vertexModels := mc.GetModelsForProvider(schemas.Vertex)
+		for _, p := range providers {
+			if slices.Contains(vertexModels, string(p)+"/"+model) {
+				providers = append(providers, schemas.Vertex)
+				seen[schemas.Vertex] = struct{}{}
+				break
+			}
+		}
+	}
+	if _, ok := seen[schemas.Groq]; !ok && strings.Contains(model, "gpt-") {
+		if slices.Contains(mc.GetModelsForProvider(schemas.Groq), "openai/"+model) {
+			providers = append(providers, schemas.Groq)
+		}
+	}
+	if _, ok := seen[schemas.Bedrock]; !ok && strings.Contains(model, "claude") {
+		for _, bedrockModel := range mc.GetModelsForProvider(schemas.Bedrock) {
+			if strings.Contains(bedrockModel, model) {
+				providers = append(providers, schemas.Bedrock)
+				break
 			}
 		}
 	}
 
-	// 3. Handle openai models for groq
-	if !slices.Contains(providers, schemas.Groq) && strings.Contains(model, "gpt-") {
-		if groqModels, ok := mc.modelPool[schemas.Groq]; ok {
-			if slices.Contains(groqModels, "openai/"+model) {
-				providers = append(providers, schemas.Groq)
-			}
+	for _, p := range mc.keyconf.Providers() {
+		if _, ok := seen[p]; ok {
+			continue
 		}
-	}
-
-	// 4. Handle anthropic models for bedrock
-	if !slices.Contains(providers, schemas.Bedrock) && strings.Contains(model, "claude") {
-		if bedrockModels, ok := mc.modelPool[schemas.Bedrock]; ok {
-			for _, bedrockModel := range bedrockModels {
-				if strings.Contains(bedrockModel, model) {
-					providers = append(providers, schemas.Bedrock)
-					break
-				}
-			}
+		if mc.keyconf.BlacklistedFor(p).IsBlocked(model) {
+			continue
+		}
+		allowed := mc.keyconf.AllowedFor(p)
+		matched := false
+		if _, hit := mc.keyconf.ResolveAlias(p, model); hit && allowed.IsAllowed(model) {
+			matched = true
+		} else if allowed.Contains(model) {
+			matched = true
+		} else if allowed.IsUnrestricted() &&
+			len(mc.datasheet.DatasheetModelsForProvider(p)) == 0 &&
+			len(mc.live.UnfilteredModelsForProvider(p)) == 0 {
+			matched = true
+		}
+		if matched {
+			providers = append(providers, p)
+			seen[p] = struct{}{}
 		}
 	}
 
 	return providers
 }
 
-// IsModelAllowedForProvider checks if a model is allowed for a specific provider
-// based on the allowed models list and catalog data. It handles all cross-provider
-// logic including provider-prefixed models and special routing rules.
+// IsModelAllowedForProvider checks whether the model is allowed for the
+// provider given an explicit allowedModels list (used by VK governance
+// checks, not by the static keyconfig allow set).
 //
-// Parameters:
-//   - provider: The provider to check against
-//   - model: The model name (without provider prefix, e.g., "gpt-4o" or "claude-3-5-sonnet")
-//   - allowedModels: List of allowed model names (can be empty, can include provider prefixes)
-//
-// Behavior:
-//   - If allowedModels is ["*"]: Uses model catalog to check if provider supports the model
-//     (delegates to GetProvidersForModel which handles all cross-provider logic)
-//   - If allowedModels is empty ([]): Deny-by-default — returns false for any provider/model pair
-//   - If allowedModels is not empty: Checks if model matches any entry in the list
-//     Provider-specific validation:
-//   - Direct matches: "gpt-4o" in allowedModels for any provider
-//   - Prefixed matches: Only if the prefixed model exists in provider's catalog
-//     (e.g., "openai/gpt-4o" in allowedModels only matches if openrouter's catalog
-//     contains "openai/gpt-4o" AND the model part matches the request)
-//
-// Returns:
-//   - bool: true if the model is allowed for the provider, false otherwise
-//
-// Examples:
-//
-//	// Wildcard allowedModels - uses catalog to check provider support
-//	mc.IsModelAllowedForProvider("openrouter", "claude-3-5-sonnet", []string{"*"})
-//	// Returns: true (catalog knows openrouter has "anthropic/claude-3-5-sonnet")
-//
-//	// Empty allowedModels - deny all (deny-by-default)
-//	mc.IsModelAllowedForProvider("openrouter", "claude-3-5-sonnet", []string{})
-//	// Returns: false (no models are permitted)
-//
-//	// Explicit allowedModels with prefix - validates against catalog
-//	mc.IsModelAllowedForProvider("openrouter", "gpt-4o", []string{"openai/gpt-4o"})
-//	// Returns: true (openrouter's catalog contains "openai/gpt-4o" AND model part is "gpt-4o")
-//
-//	// Explicit allowedModels with prefix - wrong model
-//	mc.IsModelAllowedForProvider("openrouter", "claude-3-5-sonnet", []string{"openai/gpt-4o"})
-//	// Returns: false (model part "gpt-4o" doesn't match request "claude-3-5-sonnet")
-//
-//	// Explicit allowedModels without prefix
-//	mc.IsModelAllowedForProvider("openai", "gpt-4o", []string{"gpt-4o"})
-//	// Returns: true (direct match)
+//   - allowedModels=["*"]: defer to GetProvidersForModel (with custom-provider
+//     fast path when list-models is disabled).
+//   - allowedModels=[]: deny-by-default.
+//   - explicit allowedModels: direct or provider-prefixed match against the
+//     provider's catalog.
 func (mc *ModelCatalog) IsModelAllowedForProvider(provider schemas.ModelProvider, model string, providerConfig *configstore.ProviderConfig, allowedModels schemas.WhiteList) bool {
-	// Case 1: ["*"] = allow all models; use catalog to determine support
-	// Empty allowedModels = deny all (fail-safe deny-by-default)
+	isCustomProvider := false
+	hasListModelsEndpointDisabled := false
+	if providerConfig != nil && providerConfig.CustomProviderConfig != nil {
+		isCustomProvider = true
+		hasListModelsEndpointDisabled = !providerConfig.CustomProviderConfig.IsOperationAllowed(schemas.ListModelsRequest)
+	}
+
 	if allowedModels.IsUnrestricted() {
-		// Providers whose models the catalog cannot enumerate (custom without list-models,
-		// or keyless self-hosted vLLM/Ollama/SGL) cannot be cross-checked against catalog
-		// membership, so a wildcard allow-list permits any model for them.
-		if mc.IsCatalogOpaqueProvider(provider, providerConfig) {
+		if isCustomProvider && hasListModelsEndpointDisabled {
 			return true
 		}
-		supportedProviders := mc.GetProvidersForModel(model)
-		return slices.Contains(supportedProviders, provider)
+		return slices.Contains(mc.GetProvidersForModel(model), provider)
 	}
 	if allowedModels.IsEmpty() {
 		return false
 	}
 
-	// Case 2: Explicit allowedModels = check if model matches any entry
-	// Get provider's catalog models for validation of prefixed entries
 	providerCatalogModels := mc.GetModelsForProvider(provider)
-
 	for _, allowedModel := range allowedModels {
-		// Direct match: "gpt-4o" == "gpt-4o"
 		if allowedModel == model {
 			return true
 		}
-
-		// Provider-prefixed match: verify it exists in provider's catalog first
-		// This ensures we only allow provider-specific model combinations that are actually supported
 		if strings.Contains(allowedModel, "/") {
-			// Check if this exact prefixed model exists in the provider's catalog
-			// e.g., for openrouter, check if "openai/gpt-4o" is in its catalog
 			if slices.Contains(providerCatalogModels, allowedModel) {
-				// Extract the model part and compare with request
 				_, modelPart := schemas.ParseModelString(allowedModel, "")
 				if modelPart == model {
 					return true
@@ -236,211 +233,21 @@ func (mc *ModelCatalog) IsModelAllowedForProvider(provider schemas.ModelProvider
 			}
 		}
 	}
-
 	return false
 }
 
-// IsCatalogOpaqueProvider reports whether the catalog cannot enumerate the models a provider
-// serves, so a wildcard ("*") allow-list must be honored as allow-all rather than cross-checked
-// against catalog membership. True for custom providers and for native providers the catalog has
-// no model list for (keyless self-hosted vLLM/Ollama/SGL, or providers without list-models
-// support). Shared by OSS governance and enterprise load-balancing so the rule has one definition.
-func (mc *ModelCatalog) IsCatalogOpaqueProvider(provider schemas.ModelProvider, providerConfig *configstore.ProviderConfig) bool {
-	if providerConfig != nil && providerConfig.CustomProviderConfig != nil {
-		// A custom provider is opaque only when it cannot list its models. If it supports the
-		// list-models endpoint, the catalog can enumerate its models, so it is NOT opaque.
-		return !providerConfig.CustomProviderConfig.IsOperationAllowed(schemas.ListModelsRequest)
-	}
-	if mc == nil {
-		return false
-	}
-	// Only an emptiness check is needed, so read modelPool directly under the
-	// read lock rather than calling GetModelsForProvider, which allocates and
-	// copies the full slice on this request-path check.
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
-	return len(mc.modelPool[provider]) == 0
-}
-
-// GetBaseModelName returns the canonical base model name for a given model string.
-// It uses the pre-computed base_model from the pricing catalog when available,
-// falling back to algorithmic date/version stripping for models not in the catalog.
-//
-// Examples:
-//
-//	mc.GetBaseModelName("gpt-4o")                    // Returns: "gpt-4o"
-//	mc.GetBaseModelName("openai/gpt-4o")             // Returns: "gpt-4o"
-//	mc.GetBaseModelName("gpt-4o-2024-08-06")         // Returns: "gpt-4o" (algorithmic fallback)
 func (mc *ModelCatalog) GetBaseModelName(model string) string {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
-	return mc.getBaseModelNameUnsafe(model)
+	return mc.datasheet.BaseModelName(model)
 }
 
-// getBaseModelNameUnsafe returns the canonical base model name for a given model string without locking.
-// This is used to avoid locking overhead when getting the base model name for many models.
-// Make sure the caller function is holding the read lock before calling this function.
-// It is not safe to use this function when the model pool is being updated.
-func (mc *ModelCatalog) getBaseModelNameUnsafe(model string) string {
-	// Step 1: Direct lookup in base model index
-	if base, ok := mc.baseModelIndex[model]; ok {
-		return base
-	}
-
-	// Step 2: Strip provider prefix and try again
-	_, baseName := schemas.ParseModelString(model, "")
-	if baseName != model {
-		if base, ok := mc.baseModelIndex[baseName]; ok {
-			return base
-		}
-	}
-
-	// Step 3: Fallback to algorithmic date/version stripping
-	// (for models not in the catalog, e.g., user-configured custom models)
-	return schemas.BaseModelName(baseName)
-}
-
-// IsSameModel checks if two model strings refer to the same underlying model.
-// It compares the canonical base model names derived from the pricing catalog
-// (or algorithmic fallback for models not in the catalog).
-//
-// Examples:
-//
-//	mc.IsSameModel("gpt-4o", "gpt-4o")                            // true (direct match)
-//	mc.IsSameModel("openai/gpt-4o", "gpt-4o")                     // true (same base model)
-//	mc.IsSameModel("gpt-4o", "claude-3-5-sonnet")                  // false (different models)
-//	mc.IsSameModel("openai/gpt-4o", "anthropic/claude-3-5-sonnet") // false
 func (mc *ModelCatalog) IsSameModel(model1, model2 string) bool {
-	if model1 == model2 {
-		return true
-	}
-	return mc.GetBaseModelName(model1) == mc.GetBaseModelName(model2)
+	return mc.datasheet.IsSameModel(model1, model2)
 }
 
-// DeleteModelDataForProvider deletes all model data from the pool for a given provider
-func (mc *ModelCatalog) DeleteModelDataForProvider(provider schemas.ModelProvider) {
-	mc.mu.Lock()
-	defer mc.mu.Unlock()
-
-	delete(mc.modelPool, provider)
-	delete(mc.unfilteredModelPool, provider)
-}
-
-// UpsertModelDataForProvider upserts model data for a given provider
-func (mc *ModelCatalog) UpsertModelDataForProvider(provider schemas.ModelProvider, modelData *schemas.BifrostListModelsResponse, allowedModels []schemas.Model) {
-	if modelData == nil {
-		return
-	}
-	mc.mu.Lock()
-	defer mc.mu.Unlock()
-
-	// Populating models from pricing data for the given provider
-	// Provider models map
-	providerModels := []string{}
-	// Iterate through all pricing data to collect models per provider
-	for _, pricing := range mc.pricingData {
-		// Normalize provider before adding to model pool
-		normalizedProvider := schemas.ModelProvider(normalizeProvider(pricing.Provider))
-		// We will only add models for the given provider
-		if normalizedProvider != provider {
-			continue
-		}
-		// Add model to the provider's model set (using map for deduplication)
-		if slices.Contains(providerModels, pricing.Model) {
-			continue
-		}
-		providerModels = append(providerModels, pricing.Model)
-		// Build base model index from pre-computed base_model field
-		if pricing.BaseModel != "" {
-			mc.baseModelIndex[pricing.Model] = pricing.BaseModel
-		}
-	}
-	// If modelData is empty, then we allow all models
-	if len(modelData.Data) == 0 && len(allowedModels) == 0 {
-		mc.modelPool[provider] = providerModels
-		return
-	}
-	// Here we make sure that we still keep the backup for model catalog intact
-	// So we start with a existing model pool and add the new models from incoming data
-	finalModelList := make([]string, 0)
-	seenModels := make(map[string]bool)
-	// Case where list models failed but we have allowed models from keys
-	if len(modelData.Data) == 0 && len(allowedModels) > 0 {
-		for _, allowedModel := range allowedModels {
-			parsedProvider, parsedModel := schemas.ParseModelString(allowedModel.ID, "")
-			if parsedProvider != provider {
-				continue
-			}
-			if !seenModels[parsedModel] {
-				seenModels[parsedModel] = true
-				finalModelList = append(finalModelList, parsedModel)
-			}
-		}
-	}
-	for _, model := range modelData.Data {
-		parsedProvider, parsedModel := schemas.ParseModelString(model.ID, "")
-		if parsedProvider != provider {
-			continue
-		}
-		if !seenModels[parsedModel] {
-			seenModels[parsedModel] = true
-			finalModelList = append(finalModelList, parsedModel)
-		}
-	}
-
-	if len(allowedModels) == 0 {
-		for _, model := range providerModels {
-			if !seenModels[model] {
-				seenModels[model] = true
-				finalModelList = append(finalModelList, model)
-			}
-		}
-	}
-	mc.modelPool[provider] = finalModelList
-}
-
-// UpsertUnfilteredModelDataForProvider upserts unfiltered model data for a given provider
-func (mc *ModelCatalog) UpsertUnfilteredModelDataForProvider(provider schemas.ModelProvider, modelData *schemas.BifrostListModelsResponse) {
-	if modelData == nil {
-		return
-	}
-	mc.mu.Lock()
-	defer mc.mu.Unlock()
-
-	// Populating models from pricing data for the given provider
-	providerModels := []string{}
-	seenModels := make(map[string]bool)
-	for _, pricing := range mc.pricingData {
-		normalizedProvider := schemas.ModelProvider(normalizeProvider(pricing.Provider))
-		if normalizedProvider != provider {
-			continue
-		}
-		if !seenModels[pricing.Model] {
-			seenModels[pricing.Model] = true
-			providerModels = append(providerModels, pricing.Model)
-		}
-	}
-	for _, model := range modelData.Data {
-		parsedProvider, parsedModel := schemas.ParseModelString(model.ID, "")
-		if parsedProvider != provider {
-			continue
-		}
-		if !seenModels[parsedModel] {
-			seenModels[parsedModel] = true
-			providerModels = append(providerModels, parsedModel)
-		}
-	}
-	mc.unfilteredModelPool[provider] = providerModels
-}
-
-// RefineModelForProvider refines the model for a given provider by performing a lookup
-// in mc.modelPool and using schemas.ParseModelString to extract provider and model parts.
-// e.g. "gpt-oss-120b" for groq provider -> "openai/gpt-oss-120b"
-//
-// Behavior:
-// - When the provider's catalog (mc.modelPool) yields multiple matching models, returns an error
-// - When exactly one match is found, returns the fully-qualified model (provider/model format)
-// - When the provider is not handled or no refinement is needed, returns the original model unchanged
+// RefineModelForProvider refines a model identifier for providers that need
+// a leading "provider/" segment (Groq, Replicate). Returns the original
+// model unchanged when no refinement applies, or an error when multiple
+// catalog candidates match ambiguously.
 func (mc *ModelCatalog) RefineModelForProvider(provider schemas.ModelProvider, model string) (string, error) {
 	switch provider {
 	case schemas.Groq:
@@ -454,179 +261,14 @@ func (mc *ModelCatalog) RefineModelForProvider(provider schemas.ModelProvider, m
 	return model, nil
 }
 
-// SetPricingOverrides replaces the full in-memory pricing override set.
-func (mc *ModelCatalog) SetPricingOverrides(rows []configstoreTables.TablePricingOverride) error {
-	seen := make(map[string]int, len(rows))
-	overrides := make([]PricingOverride, 0, len(rows))
-	for i := range rows {
-		o, err := convertTablePricingOverrideToPricingOverride(&rows[i])
-		if err != nil {
-			return err
-		}
-		if idx, exists := seen[o.ID]; exists {
-			overrides[idx] = o // last entry wins for duplicate IDs
-		} else {
-			seen[o.ID] = len(overrides)
-			overrides = append(overrides, o)
-		}
-	}
-	mc.overridesMu.Lock()
-	mc.rawOverrides = overrides
-	mc.customPricing = buildCustomPricingData(overrides)
-	mc.overridesMu.Unlock()
-	return nil
-}
-
-// UpsertPricingOverrides inserts or replaces one or more pricing overrides in a single
-// operation, rebuilding the lookup map only once at the end.
-func (mc *ModelCatalog) UpsertPricingOverrides(rows ...*configstoreTables.TablePricingOverride) error {
-	// Deduplicate the input batch by ID (last entry wins) and build the
-	// incoming set for O(1) lookup when filtering existing rawOverrides.
-	seenIncoming := make(map[string]int, len(rows))
-	overrides := make([]PricingOverride, 0, len(rows))
-	for _, row := range rows {
-		o, err := convertTablePricingOverrideToPricingOverride(row)
-		if err != nil {
-			return err
-		}
-		if idx, exists := seenIncoming[o.ID]; exists {
-			overrides[idx] = o // last entry wins for duplicate IDs
-		} else {
-			seenIncoming[o.ID] = len(overrides)
-			overrides = append(overrides, o)
-		}
-	}
-
-	mc.overridesMu.Lock()
-	defer mc.overridesMu.Unlock()
-
-	updated := make([]PricingOverride, 0, len(mc.rawOverrides)+len(overrides))
-	for _, o := range mc.rawOverrides {
-		if _, replacing := seenIncoming[o.ID]; !replacing {
-			updated = append(updated, o)
-		}
-	}
-	updated = append(updated, overrides...)
-	mc.rawOverrides = updated
-	mc.customPricing = buildCustomPricingData(updated)
-	return nil
-}
-
-// DeletePricingOverride removes a pricing override by ID.
-func (mc *ModelCatalog) DeletePricingOverride(id string) {
-	mc.overridesMu.Lock()
-	defer mc.overridesMu.Unlock()
-
-	updated := make([]PricingOverride, 0, len(mc.rawOverrides))
-	for _, o := range mc.rawOverrides {
-		if o.ID != id {
-			updated = append(updated, o)
-		}
-	}
-	mc.rawOverrides = updated
-	mc.customPricing = buildCustomPricingData(updated)
-}
-
-// IsTextCompletionSupported checks if a model supports text completion for the given provider.
-// Returns true if the model has pricing data for text completion ("text_completion"),
-// false otherwise. This is used by the litellmcompat plugin to determine whether to
-// convert text completion requests to chat completion requests.
-func (mc *ModelCatalog) IsTextCompletionSupported(model string, provider schemas.ModelProvider) bool {
-	mc.mu.RLock()
-	defer mc.mu.RUnlock()
-	// Check for text completion mode in pricing data
-	key := makeKey(model, normalizeProvider(string(provider)), normalizeRequestType(schemas.TextCompletionRequest))
-	_, ok := mc.pricingData[key]
-	return ok
-}
-
-// HELPER FUNCTIONS
-
-func (mc *ModelCatalog) getCapabilityEntryForExactModelUnsafe(model string, provider schemas.ModelProvider) *PricingEntry {
-	preferredModes := []schemas.RequestType{
-		schemas.ChatCompletionRequest,
-		schemas.ResponsesRequest,
-		schemas.TextCompletionRequest,
-	}
-
-	for _, mode := range preferredModes {
-		key := makeKey(model, string(provider), normalizeRequestType(mode))
-		pricing, ok := mc.pricingData[key]
-		if ok {
-			return convertTableModelPricingToPricingData(&pricing)
-		}
-	}
-
-	prefix := model + "|" + string(provider) + "|"
-	matchingKeys := make([]string, 0)
-	for key := range mc.pricingData {
-		if strings.HasPrefix(key, prefix) {
-			matchingKeys = append(matchingKeys, key)
-		}
-	}
-	return mc.selectCapabilityEntryFromKeysUnsafe(matchingKeys)
-}
-
-func (mc *ModelCatalog) getCapabilityEntryForModelFamilyUnsafe(baseModel string, provider schemas.ModelProvider) *PricingEntry {
-	if baseModel == "" {
-		return nil
-	}
-
-	matchingKeys := make([]string, 0)
-	for key, pricing := range mc.pricingData {
-		if normalizeProvider(pricing.Provider) != string(provider) {
-			continue
-		}
-		if mc.getBaseModelNameUnsafe(pricing.Model) != baseModel {
-			continue
-		}
-		matchingKeys = append(matchingKeys, key)
-	}
-	return mc.selectCapabilityEntryFromKeysUnsafe(matchingKeys)
-}
-
-func (mc *ModelCatalog) selectCapabilityEntryFromKeysUnsafe(matchingKeys []string) *PricingEntry {
-	if len(matchingKeys) == 0 {
-		return nil
-	}
-
-	preferredModes := []string{
-		normalizeRequestType(schemas.ChatCompletionRequest),
-		normalizeRequestType(schemas.ResponsesRequest),
-		normalizeRequestType(schemas.TextCompletionRequest),
-	}
-
-	for _, mode := range preferredModes {
-		modeMatches := make([]string, 0)
-		for _, key := range matchingKeys {
-			parts := strings.SplitN(key, "|", 3)
-			if len(parts) != 3 || parts[2] != mode {
-				continue
-			}
-			modeMatches = append(modeMatches, key)
-		}
-		if len(modeMatches) == 0 {
-			continue
-		}
-		slices.Sort(modeMatches)
-		pricing := mc.pricingData[modeMatches[0]]
-		return convertTableModelPricingToPricingData(&pricing)
-	}
-
-	slices.Sort(matchingKeys)
-	pricing := mc.pricingData[matchingKeys[0]]
-	return convertTableModelPricingToPricingData(&pricing)
-}
-
 // refineNestedProviderModel resolves provider-native model slugs such as
-// "openai/gpt-5-nano" from a base model request like "gpt-5-nano".
-// It only considers catalog entries whose leading segment is a known Bifrost provider,
-// so Replicate owner/model identifiers like "meta/llama-3-8b" are left untouched.
+// "openai/gpt-5-nano" from a base model request like "gpt-5-nano". Only
+// considers catalog entries whose leading segment is a known Bifrost
+// provider so Replicate owner/model identifiers like "meta/llama-3-8b" are
+// left untouched.
 func (mc *ModelCatalog) refineNestedProviderModel(provider schemas.ModelProvider, model string) (string, error) {
-	mc.mu.RLock()
-	models, ok := mc.modelPool[provider]
-	mc.mu.RUnlock()
-	if !ok {
+	models := mc.GetModelsForProvider(provider)
+	if len(models) == 0 {
 		return model, nil
 	}
 
@@ -637,7 +279,6 @@ func (mc *ModelCatalog) refineNestedProviderModel(provider schemas.ModelProvider
 		if providerPart == "" || model != modelPart {
 			continue
 		}
-
 		candidate := string(providerPart) + "/" + modelPart
 		if _, seen := seenCandidates[candidate]; seen {
 			continue
